Refactor refineCollectedData method to a separate class. (#339)
Spreadsheet standardization and refining collected data (e.g. getting to
the JSON data we serve) were in the same class for historical reasons,
but these processes are pretty separate now.

#namingishard and if you have a better name than DataRefiner, happy to
change it.
