+++ test/scripts/incentive-spreadsheet-to-json.test.ts
@@ -1,5 +1,6 @@
 import { test } from 'tap';
 import { spreadsheetToJson } from '../../scripts/incentive-spreadsheet-to-json';
+import { DataRefiner } from '../../scripts/lib/data-refiner';
 import { flatToNestedValidate } from '../../scripts/lib/format-converter';
 import {
   FIELD_MAPPINGS,
@@ -89,13 +90,13 @@ test('correct row to record transformation', tap => {
     FIELD_MAPPINGS,
     VALUE_MAPPINGS,
     true,
-    fakeIncomeThresholds,
   );
+  const refiner = new DataRefiner(fakeIncomeThresholds);
   for (const tc of testCases) {
     const standardized = standardizer.standardize(tc.input);
     const [valids, invalids] = flatToNestedValidate([standardized]);
     tap.equal(invalids.length, 0, `Invalid record: ${invalids[0]}`);
-    tap.matchOnly(standardizer.refineCollectedData('va', valids[0]), tc.want);
+    tap.matchOnly(refiner.refineCollectedData('va', valids[0]), tc.want);
   }
   tap.end();
 });
